refactor: rename optimization.py to optimization_metrics_vector.py for naming consistency
