+++ analysis/utils/analysis_utils.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Tuple
 
 import pandas as pd
 
-from application.services.pair_generation.optimization import (
+from application.services.pair_generation.optimization_metrics_vector import (
     OptimizationMetricsStrategy,
 )
