+++ application/services/pair_generation/optimization_metrics_vector.py
